requirements.txt
2to3==1.0
absl-py==2.0.0
anyio==3.7.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
gunicorn==19.9.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
caer==2.0.8
catboost==1.2.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.1
djangorestframework==3.14.0
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.17.1
feedparser==6.0.10
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
future==0.18.2
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gpg==1.16.0
graphviz==0.20.1
grpcio==1.60.0
gyp==0.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.1.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.25.2
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lightgbm==4.3.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.2.1
mypy==1.3.0
mypy-extensions==1.0.0
natsort==8.0.2
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.1
nbformat==5.9.1
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
notebook==7.1.0
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.7.0.72
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.9.1
playsound==1.3.0
plotly==5.19.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydot==2.0.0
pygame==2.3.0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
sutree==0.2.0
sympy==1.11.1
systemd-python==234
tartube==2.4.370
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.6.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.180
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xgboost==2.0.3
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
